routes and services created, added bookmark icon to recipe cards and suggestion cards
