+++ api/routes/recipes.js
@@ -2,10 +2,8 @@ const express = require("express");
 const router = express.Router();
 const RecipesController = require("../controllers/recipes");
 
-
 router.get("/", RecipesController.getRandomRecipes);
-
-// router.post("/", RecipesController.);
-
+router.post("/savedrecipes", RecipesController.addRecipeToFavourites);
+router.delete("/savedrecipes", RecipesController.removeRecipeFromFavourites)
 
 module.exports = router;
